UploadCameraImageEdgeToCloud/modules/StoreImageToLocalBlob/requirements.txt
azure-iothub-device-client
flask
azure-storage-blob
azure-mgmt-storage
